feat: update Layout and nested layout to copy Outlet functionality
